Corrected notifications/debugging
Now using the correct vim.notify format;
Also changed the levels. No bib specification in document leads to a
DEBUG level messsage. This is because the user may simply not have a bib
for that file! If however a bib specification is found then cannot be
read that will flag a warning.